Forgot to also include the .h changes

+++ src/motor_driver.h
@@ -46,7 +46,7 @@ class MotorDriver : public rclcpp::Node {
   bool publish_joint_states_;
   std::thread publisher_thread_;  // For publishing odom, joint state, etc.
   uint32_t quad_pulses_per_meter_;
-  uint32_t quad_pulses_per_revolution_;
+  float quad_pulses_per_revolution_;
   uint8_t vmin_;
   uint8_t vtime_;
   double wheel_radius_;
